+++ src/main/java/heyso/HeysoDiaryBackEnd/diaryAi/dto/DiaryAiCommentCreateRequest.java
@@ -1,7 +1,5 @@
 package heyso.HeysoDiaryBackEnd.diaryAi.dto;
 
-import java.math.BigDecimal;
-
 import jakarta.validation.constraints.DecimalMax;
 import jakarta.validation.constraints.DecimalMin;
 import jakarta.validation.constraints.Positive;
@@ -25,11 +23,11 @@ public class DiaryAiCommentCreateRequest {
 
     @DecimalMin("0.0")
     @DecimalMax("2.0")
-    private BigDecimal temperature = BigDecimal.valueOf(0.7);
+    private Double temperature = Double.valueOf(0.7);
 
     @DecimalMin("0.0")
     @DecimalMax("1.0")
-    private BigDecimal topP = BigDecimal.valueOf(0.9);
+    private Double topP = Double.valueOf(0.9);
 
     @Positive
     private Integer maxOutputTokens = 1000;
